requirements.txt
beautifulsoup4==4.6.0
certifi==2018.4.16
chardet==3.0.4
future==0.16.0
idna==2.6
mysql-connector-python==8.0.11
oauthlib==2.0.7
protobuf==3.5.2.post1
python-twitter==3.4.1
requests==2.18.4
requests-oauthlib==0.8.0
six==1.11.0
urllib3==1.22
